apply mvvm to email screen
